requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
ecapture==2.0.2
executing==1.2.0
filelock==3.9.0
Flask==2.2.2
idna==3.4
ipython==8.10.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
more-itertools==9.0.0
numpy==1.24.2
opencv-python==4.7.0.68
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipdeptree==2.3.3
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
PyAudio==0.2.13
Pygments==2.14.0
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7.1
pywin32==305
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
source==1.2.0
SpeechRecognition==3.9.0
stack-data==0.6.2
torch==1.13.1
torchvision==0.14.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.17.1
wcwidth==0.2.6
Werkzeug==2.2.2
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
yarg==0.1.9
